refactor(gfn): split errorCodes and keyboard scancode tables
Move GfnErrorCode and ERROR_MESSAGES into dedicated modules, and extract
static keyboard scancode/char maps from inputProtocol for easier maintenance.
