+++ opennow-stable/src/renderer/src/platforms/gfn/keyboardScancodes.ts
@@ -1,0 +1,298 @@
+/** Static keyboard scancode and character-code maps for GFN input protocol. */
+
+interface KeyMapping {
+  vk: number;
+  scancode: number;
+}
+
+export const scancodeByCode: Record<string, number> = {
+  KeyA: 0x001e,
+  KeyB: 0x0030,
+  KeyC: 0x002e,
+  KeyD: 0x0020,
+  KeyE: 0x0012,
+  KeyF: 0x0021,
+  KeyG: 0x0022,
+  KeyH: 0x0023,
+  KeyI: 0x0017,
+  KeyJ: 0x0024,
+  KeyK: 0x0025,
+  KeyL: 0x0026,
+  KeyM: 0x0032,
+  KeyN: 0x0031,
+  KeyO: 0x0018,
+  KeyP: 0x0019,
+  KeyQ: 0x0010,
+  KeyR: 0x0013,
+  KeyS: 0x001f,
+  KeyT: 0x0014,
+  KeyU: 0x0016,
+  KeyV: 0x002f,
+  KeyW: 0x0011,
+  KeyX: 0x002d,
+  KeyY: 0x0015,
+  KeyZ: 0x002c,
+  Digit1: 0x0002,
+  Digit2: 0x0003,
+  Digit3: 0x0004,
+  Digit4: 0x0005,
+  Digit5: 0x0006,
+  Digit6: 0x0007,
+  Digit7: 0x0008,
+  Digit8: 0x0009,
+  Digit9: 0x000a,
+  Digit0: 0x000b,
+  Enter: 0x001c,
+  Escape: 0x0001,
+  Backspace: 0x000e,
+  Tab: 0x000f,
+  Space: 0x0039,
+  Minus: 0x000c,
+  Equal: 0x000d,
+  BracketLeft: 0x001a,
+  BracketRight: 0x001b,
+  Backslash: 0x002b,
+  IntlBackslash: 0x0056,
+  IntlRo: 0x0073,
+  IntlYen: 0x007d,
+  Semicolon: 0x0027,
+  Quote: 0x0028,
+  Backquote: 0x0029,
+  Comma: 0x0033,
+  Period: 0x0034,
+  Slash: 0x0035,
+  F1: 0x003b,
+  F2: 0x003c,
+  F3: 0x003d,
+  F4: 0x003e,
+  F5: 0x003f,
+  F6: 0x0040,
+  F7: 0x0041,
+  F8: 0x0042,
+  F9: 0x0043,
+  F10: 0x0044,
+  F11: 0x0057,
+  F12: 0x0058,
+  F13: 0x0064,
+  ArrowRight: 0xe04d,
+  ArrowLeft: 0xe04b,
+  ArrowDown: 0xe050,
+  ArrowUp: 0xe048,
+  ControlLeft: 0x001d,
+  ShiftLeft: 0x002a,
+  AltLeft: 0x0038,
+  MetaLeft: 0xe05b,
+  ControlRight: 0xe01d,
+  ShiftRight: 0x0036,
+  AltRight: 0xe038,
+  MetaRight: 0xe05c,
+  CapsLock: 0x003a,
+  NumLock: 0xe045,
+  Insert: 0xe052,
+  Delete: 0xe053,
+  Home: 0xe047,
+  End: 0xe04f,
+  PageUp: 0xe049,
+  PageDown: 0xe051,
+  PrintScreen: 0xe037,
+  ScrollLock: 0x0046,
+  Pause: 0x0045,
+  ContextMenu: 0xe05d,
+  Numpad0: 0x0052,
+  Numpad1: 0x004f,
+  Numpad2: 0x0050,
+  Numpad3: 0x0051,
+  Numpad4: 0x004b,
+  Numpad5: 0x004c,
+  Numpad6: 0x004d,
+  Numpad7: 0x0047,
+  Numpad8: 0x0048,
+  Numpad9: 0x0049,
+  NumpadAdd: 0x004e,
+  NumpadSubtract: 0x004a,
+  NumpadMultiply: 0x0037,
+  NumpadDivide: 0xe035,
+  NumpadDecimal: 0x0053,
+  NumpadEnter: 0xe01c,
+  NumpadEqual: 0x0059,
+  NumpadComma: 0x007e,
+};
+
+export const specialVirtualKeyByCode: Record<string, number> = {
+  Enter: 0x0d,
+  Escape: 0x1b,
+  Backspace: 0x08,
+  Tab: 0x09,
+  Space: 0x20,
+  Minus: 0xbd,
+  Equal: 0xbb,
+  BracketLeft: 0xdb,
+  BracketRight: 0xdd,
+  Backslash: 0xdc,
+  IntlBackslash: 0xe2,
+  IntlRo: 0xc2,
+  IntlYen: 0xc1,
+  Semicolon: 0xba,
+  Quote: 0xde,
+  Backquote: 0xc0,
+  Comma: 0xbc,
+  Period: 0xbe,
+  Slash: 0xbf,
+  ArrowRight: 0x27,
+  ArrowLeft: 0x25,
+  ArrowDown: 0x28,
+  ArrowUp: 0x26,
+  ControlLeft: 0xa2,
+  ShiftLeft: 0xa0,
+  AltLeft: 0xa4,
+  MetaLeft: 0x5b,
+  ControlRight: 0xa3,
+  ShiftRight: 0xa1,
+  AltRight: 0xa5,
+  MetaRight: 0x5c,
+  CapsLock: 0x14,
+  NumLock: 0x90,
+  Insert: 0x2d,
+  Delete: 0x2e,
+  Home: 0x24,
+  End: 0x23,
+  PageUp: 0x21,
+  PageDown: 0x22,
+  PrintScreen: 0x2a,
+  ScrollLock: 0x91,
+  Pause: 0x13,
+  ContextMenu: 0x5d,
+  OSLeft: 0x5b,
+  OSRight: 0x5c,
+  KanaMode: 0xe9,
+  Lang1: 0x15,
+  Lang2: 0x19,
+  Convert: 0xea,
+  NonConvert: 0xeb,
+  NumpadClear: 0x0c,
+  NumpadClearEntry: 0x0c,
+  NumpadAdd: 0x6b,
+  NumpadSubtract: 0x6d,
+  NumpadMultiply: 0x6a,
+  NumpadDivide: 0x6f,
+  NumpadDecimal: 0x6e,
+  NumpadEnter: 0x0d,
+  NumpadEqual: 0xbb,
+  NumpadComma: 0xbc,
+};
+
+export const keyFallbackMap: Record<string, KeyMapping> = {
+  Escape: { vk: 0x1b, scancode: 0x0001 },
+  Esc: { vk: 0x1b, scancode: 0x0001 },
+};
+
+export const baseCharCodeMap: Record<string, string> = {
+  " ": "Space",
+  "\n": "Enter",
+  "\r": "Enter",
+  "\t": "Tab",
+  "0": "Digit0",
+  "1": "Digit1",
+  "2": "Digit2",
+  "3": "Digit3",
+  "4": "Digit4",
+  "5": "Digit5",
+  "6": "Digit6",
+  "7": "Digit7",
+  "8": "Digit8",
+  "9": "Digit9",
+  "-": "Minus",
+  "=": "Equal",
+  "[": "BracketLeft",
+  "]": "BracketRight",
+  "\\": "Backslash",
+  ";": "Semicolon",
+  "'": "Quote",
+  "`": "Backquote",
+  ",": "Comma",
+  ".": "Period",
+  "/": "Slash",
+};
+
+export const shiftedCharCodeMap: Record<string, string> = {
+  "!": "Digit1",
+  "@": "Digit2",
+  "#": "Digit3",
+  "$": "Digit4",
+  "%": "Digit5",
+  "^": "Digit6",
+  "&": "Digit7",
+  "*": "Digit8",
+  "(": "Digit9",
+  ")": "Digit0",
+  "_": "Minus",
+  "+": "Equal",
+  "{": "BracketLeft",
+  "}": "BracketRight",
+  "|": "Backslash",
+  ":": "Semicolon",
+  '"': "Quote",
+  "~": "Backquote",
+  "<": "Comma",
+  ">": "Period",
+  "?": "Slash",
+};
+
+export const germanBaseCharCodeMap: Record<string, string> = {
+  " ": "Space",
+  "\n": "Enter",
+  "\r": "Enter",
+  "\t": "Tab",
+  "0": "Digit0",
+  "1": "Digit1",
+  "2": "Digit2",
+  "3": "Digit3",
+  "4": "Digit4",
+  "5": "Digit5",
+  "6": "Digit6",
+  "7": "Digit7",
+  "8": "Digit8",
+  "9": "Digit9",
+  "y": "KeyZ",
+  "z": "KeyY",
+  "ß": "Minus",
+  "´": "Equal",
+  "ü": "BracketLeft",
+  "+": "BracketRight",
+  "#": "Backslash",
+  "ö": "Semicolon",
+  "ä": "Quote",
+  ",": "Comma",
+  ".": "Period",
+  "^": "Backquote",
+  "-": "Slash",
+  "<": "IntlBackslash",
+};
+
+export const germanShiftedCharCodeMap: Record<string, string> = {
+  "!": "Digit1",
+  '"': "Digit2",
+  "§": "Digit3",
+  "$": "Digit4",
+  "%": "Digit5",
+  "&": "Digit6",
+  "/": "Digit7",
+  "(": "Digit8",
+  ")": "Digit9",
+  "=": "Digit0",
+  "Y": "KeyZ",
+  "Z": "KeyY",
+  "?": "Minus",
+  "`": "Equal",
+  "Ü": "BracketLeft",
+  "*": "BracketRight",
+  "'": "Backslash",
+  "Ö": "Semicolon",
+  "Ä": "Quote",
+  "°": "Backquote",
+  ";": "Comma",
+  ":": "Period",
+  "_": "Slash",
+  ">": "IntlBackslash",
+};
+
